Added keyboard control of the mouse for menus
Go to music player (should be renamed to a manager or something) and tick menucursor to true and you can move the mouse with arrow keys or right stick. E/a(south) to select option or esc/b(west) to go back. (Main menu works properly, need to change things in scene to make it work in other levels)